backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Cache and Background Tasks
redis==5.0.1
celery==5.3.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0

# HTTP Requests
httpx==0.25.2
aiohttp==3.9.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
pytz==2023.3

# AI and External APIs
google-generativeai==0.3.2
openai==1.3.7  # For backup AI services

# Text-to-Speech
gTTS==2.4.0
edge-tts==6.1.9

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0